+++ proyecto_Integrador/src/main/java/com/example/proyecto_Integrador/entity/Domicilio.java
@@ -8,7 +8,7 @@ public class Domicilio {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @Column
     private String calle;
     @Column
@@ -21,7 +21,7 @@ public class Domicilio {
     public Domicilio() {
     }
 
-    public Domicilio(Integer id, String calle, Integer numero, String localidad, String provincia) {
+    public Domicilio(Long id, String calle, Integer numero, String localidad, String provincia) {
         this.id = id;
         this.calle = calle;
         this.numero = numero;
@@ -36,11 +36,11 @@ public class Domicilio {
         this.provincia = provincia;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
