+++ proyecto_Integrador/src/main/java/com/example/proyecto_Integrador/entity/Turno.java
@@ -9,7 +9,7 @@ public class Turno {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @OneToOne(cascade = CascadeType.ALL)
     @JoinColumn(name = "paciente_id")
     private Paciente paciente;
@@ -22,7 +22,7 @@ public class Turno {
     public Turno() {
     }
 
-    public Turno(Integer id, Paciente paciente, Odontologo odontologo, LocalDate fecha) {
+    public Turno(Long id, Paciente paciente, Odontologo odontologo, LocalDate fecha) {
         this.id = id;
         this.paciente = paciente;
         this.odontologo = odontologo;
@@ -35,11 +35,11 @@ public class Turno {
         this.fecha = fecha;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
