+++ proyecto_Integrador/src/main/java/com/example/proyecto_Integrador/repository/DomicilioRepository.java
@@ -3,7 +3,7 @@ package com.example.proyecto_Integrador.repository;
 import com.example.proyecto_Integrador.entity.Domicilio;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DomicilioRepository extends JpaRepository<Domicilio, Integer> {
+public interface DomicilioRepository extends JpaRepository<Domicilio, Long> {
 
 
 }
