+++ proyecto_Integrador/src/main/java/com/example/proyecto_Integrador/repository/TurnoRepository.java
@@ -3,5 +3,5 @@ package com.example.proyecto_Integrador.repository;
 import com.example.proyecto_Integrador.entity.Turno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TurnoRepository extends JpaRepository<Turno, Integer> {
+public interface TurnoRepository extends JpaRepository<Turno, Long> {
 }
